refactor(core): extract initiative family owners
