+++ framework/core/src/python/kungfu/assignment_orchestration.py
@@ -4,7 +4,6 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import ntpath
 import os
@@ -16,7 +15,10 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any, Mapping
 
-ROOT = "sha256:"
+from kungfu import initiative_family
+from kungfu.initiative_family import canonical as assignment_canonical
+from kungfu.initiative_family import typed_v2 as initiative_family_v2
+
 REQUEST_SCHEMA = "kungfu.assignment-request/v1"
 CAPTURE_RECEIPT_SCHEMA = "kungfu.assignment-capture.receipt/v1"
 CAPTURE_RESPONSE_SCHEMA = "kungfu.assignment-capture.response/v1"
@@ -29,24 +31,11 @@ OUTCOME_BINDING_SCHEMA = (
     "kungfu.assignment-orchestration.work-design-outcome-binding/v1"
 )
 OUTCOME_INDEX_SCHEMA = "kungfu.assignment-orchestration.work-design-outcome-index/v1"
-FAMILY_BLUEPRINT_SCHEMA = "kungfu.work-control.initiative-family-blueprint/v1"
-FAMILY_STATE_SCHEMA = "kungfu.work-control.initiative-family-state/v1"
-FAMILY_TRANSITION_SCHEMA = "kungfu.work-control.initiative-family-transition/v1"
-FAMILY_CONTRACT_V2_SCHEMA = "kungfu.work-control.initiative-family-contract/v2"
-FAMILY_STATE_V2_SCHEMA = "kungfu.work-control.initiative-family-state/v2"
-FAMILY_TRANSITION_V2_SCHEMA = "kungfu.work-control.initiative-family-transition/v2"
-FAMILY_BINDING_V2_SCHEMA = (
-    "kungfu.work-control.initiative-family-typed-binding-manifest/v2"
-)
-FAMILY_UPGRADE_V2_SCHEMA = "kungfu.work-control.initiative-family-upgrade/v2"
 CROSS_WORKSPACE_BINDING_SCHEMA = (
     "kungfu.assignment-orchestration.cross-workspace-binding/v1"
 )
 PRODUCT_MANIFEST_SCHEMA = "kungfu.product-upgrade.manifest/v1"
 _GIT_REVISION = re.compile(r"^[0-9a-f]{40}$")
-_ISO_8601 = re.compile(
-    r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?(?:Z|[+-]\d{2}:\d{2})$"
-)
 PHASES = (
     "admitted",
     "claimed",
@@ -56,1066 +45,43 @@ PHASES = (
     "independently-reviewed",
     "continuation-decided",
 )
-FAMILY_DELIVERY_CLASSES = (
-    "non-native-fast",
-    "native-proof-required",
-    "cross-platform",
-    "release",
-)
-FAMILY_TERMINAL_STATES = ("merged", "continued", "deferred", "failed")
-FAMILY_ACCEPTANCE_STATES = ("proved", "partial", "missing", "invalidated")
-FAMILY_V2_REFERENCE_KINDS = (
-    "acceptance-policy",
-    "admission-receipt",
-    "assessment",
-    "assignment-state",
-    "completion-claim",
-    "decision",
-    "delivery-evidence",
-    "episode",
-    "execution-warrant",
-    "atlas",
-    "project-cut",
-    "publication-failure",
-    "pursuit",
-    "work-definition",
-)
-
-
-def _normalized(value: Any) -> Any:
-    if isinstance(value, str):
-        return unicodedata.normalize("NFC", value)
-    if isinstance(value, list):
-        return [_normalized(row) for row in value]
-    if isinstance(value, dict):
-        return {
-            unicodedata.normalize("NFC", str(key)): _normalized(item)
-            for key, item in value.items()
-        }
-    if value is None or isinstance(value, (bool, int)):
-        return value
-    raise ValueError(f"unsupported canonical JSON value: {type(value).__name__}")
-
-
-def canonical_json(value: Any) -> str:
-    return json.dumps(
-        _normalized(value),
-        ensure_ascii=False,
-        separators=(",", ":"),
-        sort_keys=True,
-    )
-
-
-def semantic_root(value: Any) -> str:
-    return ROOT + hashlib.sha256(canonical_json(value).encode("utf-8")).hexdigest()
-
-
-def _strict_object(value: Any, fields: set[str], label: str) -> dict[str, Any]:
-    if not isinstance(value, dict) or set(value) != fields:
-        raise ValueError(f"{label} has an invalid field set")
-    return value
-
-
-def _sorted_unique_strings(
-    value: Any, field: str, *, allow_empty: bool = False
-) -> list[str]:
-    if (
-        not isinstance(value, list)
-        or not all(isinstance(row, str) and row for row in value)
-        or value != sorted(set(value), key=lambda row: row.encode("utf-8"))
-        or (not allow_empty and not value)
-    ):
-        qualifier = "" if allow_empty else " non-empty"
-        raise ValueError(f"{field} must be a sorted unique{qualifier} string array")
-    return value
-
-
-def _timestamp(value: Any, field: str) -> datetime:
-    if not isinstance(value, str) or not _ISO_8601.fullmatch(value):
-        raise ValueError(f"{field} must be an ISO-8601 timestamp")
-    try:
-        parsed = datetime.fromisoformat(value.replace("Z", "+00:00"))
-    except ValueError as error:
-        raise ValueError(f"{field} must be an ISO-8601 timestamp") from error
-    if parsed.tzinfo is None:
-        raise ValueError(f"{field} must include a timezone")
-    return parsed
-
-
-def _family_state_root(value: Mapping[str, Any]) -> str:
-    preimage = dict(value)
-    preimage.pop("stateRoot", None)
-    return semantic_root(preimage)
-
-
-def family_contract() -> dict[str, Any]:
-    """Return the versioned native Initiative-family protocol surface."""
-
-    body = {
-        "schema": "kungfu.work-control.initiative-family-contract/v1",
-        "blueprintSchema": FAMILY_BLUEPRINT_SCHEMA,
-        "stateSchema": FAMILY_STATE_SCHEMA,
-        "transitionSchema": FAMILY_TRANSITION_SCHEMA,
-        "deliveryClasses": list(FAMILY_DELIVERY_CLASSES),
-        "terminalStates": list(FAMILY_TERMINAL_STATES),
-        "acceptanceStates": list(FAMILY_ACCEPTANCE_STATES),
-        "waveChildBounds": {"minimum": 3, "maximum": 6},
-        "continuationDeadlineSeconds": 300,
-        "authority": {
-            "initiativeParent": "inert",
-            "waveGate": "membership-only-terminal",
-            "assignment": "bounded-execution-unit",
-        },
-        "parentDeniedAuthorities": [
-            "execution-claim",
-            "execution-lease",
-            "task-worktree",
-            "code-pull-request",
-            "merge-queue-lease",
-        ],
-    }
-    return {**body, "contractRoot": semantic_root(body)}
-
-
-def family_contract_v2() -> dict[str, Any]:
-    """Return the additive typed-envelope contract without changing v1."""
-
-    body = {
-        "schema": FAMILY_CONTRACT_V2_SCHEMA,
-        "predecessorContractRoot": family_contract()["contractRoot"],
-        "stateSchema": FAMILY_STATE_V2_SCHEMA,
-        "transitionSchema": FAMILY_TRANSITION_V2_SCHEMA,
-        "typedBindingSchema": FAMILY_BINDING_V2_SCHEMA,
-        "upgradeSchema": FAMILY_UPGRADE_V2_SCHEMA,
-        "typedReferenceKinds": list(FAMILY_V2_REFERENCE_KINDS),
-        "typedReferenceFields": [
-            "cutRoot",
-            "factWorld",
-            "identity",
-            "kind",
-            "root",
-            "schema",
-            "status",
-        ],
-        "authority": {
-            "familyState": "coordination-projection-only",
-            "initiativeParent": "inert",
-            "referencedAuthorities": "retained-by-owner",
-            "upgradeBindings": "caller-supplied-no-inference",
-        },
-        "compatibility": {
-            "v1ReaderState": "under-typed",
-            "v1Projection": "exact-root-preserving",
-            "rewritesExistingRoots": False,
-        },
-    }
-    return {**body, "contractRoot": semantic_root(body)}
-
-
-def _validate_family_terminal(
-    terminal: Any, child: Mapping[str, Any]
-) -> dict[str, Any]:
-    if not isinstance(terminal, dict):
-        raise ValueError("child.terminal must be null or an object")
-    state = str(terminal.get("state") or "")
-    field_sets = {
-        "merged": {
-            "state",
-            "recordedAt",
-            "sourceRoot",
-            "pullRequestRoot",
-            "mergeCommitRoot",
-            "finalAncestryRoot",
-            "proofRoot",
-            "sloRoot",
-        },
-        "continued": {
-            "state",
-            "recordedAt",
-            "boundExceededAt",
-            "sourceRoot",
-            "decisionRoot",
-            "completedEvidenceRoots",
-            "completedResponsibilitySlices",
-            "residualSuccessor",
-        },
-        "deferred": {"state", "recordedAt", "sourceRoot", "decisionRoot"},
-        "failed": {"state", "recordedAt", "sourceRoot", "failureRoot"},
-    }
-    if state not in field_sets:
-        raise ValueError("child terminal state is unsupported")
-    _strict_object(terminal, field_sets[state], "child.terminal")
-    recorded_at = _timestamp(terminal["recordedAt"], "terminal.recordedAt")
-    for field, value in terminal.items():
-        if field.endswith("Root"):
-            _root(value, f"terminal.{field}")
-    if state != "continued":
-        return terminal
-
-    exceeded_at = _timestamp(terminal["boundExceededAt"], "terminal.boundExceededAt")
-    elapsed = (recorded_at - exceeded_at).total_seconds()
-    if elapsed < 0 or elapsed > 300:
-        raise ValueError(
-            "continued terminal state must be recorded within five minutes"
-        )
-    completed_evidence = _sorted_unique_strings(
-        terminal["completedEvidenceRoots"], "terminal.completedEvidenceRoots"
-    )
-    for root in completed_evidence:
-        _root(root, "terminal.completedEvidenceRoots")
-    completed = _sorted_unique_strings(
-        terminal["completedResponsibilitySlices"],
-        "terminal.completedResponsibilitySlices",
-    )
-    responsibility = list(child["responsibilitySlices"])
-    if any(row not in responsibility for row in completed):
-        raise ValueError("completed responsibility is outside the original child scope")
-    residual = [row for row in responsibility if row not in completed]
-    if not residual:
-        raise ValueError("continued terminal state requires residual responsibility")
-    successor = _strict_object(
-        terminal["residualSuccessor"],
-        {
-            "assignmentId",
-            "requestRoot",
-            "captureReceiptRoots",
-            "responsibilitySlices",
-        },
-        "terminal.residualSuccessor",
-    )
-    if (
-        not isinstance(successor["assignmentId"], str)
-        or not successor["assignmentId"]
-        or successor["assignmentId"] == child["assignmentId"]
-    ):
-        raise ValueError("residual successor must have one new Assignment identity")
-    _root(successor["requestRoot"], "terminal.residualSuccessor.requestRoot")
-    captures = _sorted_unique_strings(
-        successor["captureReceiptRoots"],
-        "terminal.residualSuccessor.captureReceiptRoots",
-    )
-    for root in captures:
-        _root(root, "terminal.residualSuccessor.captureReceiptRoots")
-    successor_slices = _sorted_unique_strings(
-        successor["responsibilitySlices"],
-        "terminal.residualSuccessor.responsibilitySlices",
-    )
-    if successor_slices != residual:
-        raise ValueError(
-            "residual successor must contain exactly the uncompleted responsibility"
-        )
-    return terminal
-
-
-def validate_family_state(value: Any) -> dict[str, Any]:
-    """Validate one immutable Initiative-family state and its semantic root."""
-
-    state = _strict_object(
-        value,
-        {
-            "schema",
-            "initiative",
-            "wave",
-            "children",
-            "acceptance",
-            "previousStateRoot",
-            "stateRoot",
-        },
-        "Initiative-family state",
-    )
-    if state["schema"] != FAMILY_STATE_SCHEMA:
-        raise ValueError(f"family state schema must be {FAMILY_STATE_SCHEMA}")
-    previous = str(state["previousStateRoot"] or "")
-    if previous:
-        _root(previous, "previousStateRoot")
-    initiative = _strict_object(
-        state["initiative"],
-        {"initiativeId", "versionRoot", "role"},
-        "initiative",
-    )
-    if (
-        not isinstance(initiative["initiativeId"], str)
-        or not initiative["initiativeId"]
-    ):
-        raise ValueError("initiative.initiativeId must be a non-empty string")
-    _root(initiative["versionRoot"], "initiative.versionRoot")
-    if initiative["role"] != "inert-parent":
-        raise ValueError("Initiative family parent must remain inert")
-    wave = _strict_object(
-        state["wave"],
-        {"waveId", "ordinal", "gateAssignmentId", "gateState"},
-        "wave",
-    )
-    if not all(
-        isinstance(wave[field], str) and wave[field]
-        for field in ("waveId", "gateAssignmentId")
-    ):
-        raise ValueError("Wave and Gate Assignment identities are required")
-    if (
-        not isinstance(wave["ordinal"], int)
-        or isinstance(wave["ordinal"], bool)
-        or wave["ordinal"] < 0
-    ):
-        raise ValueError("wave.ordinal must be a non-negative integer")
-    if wave["gateState"] != "terminal":
-        raise ValueError("Wave Gate must terminate after publishing membership")
-
-    children = state["children"]
-    if not isinstance(children, list) or not 3 <= len(children) <= 6:
-        raise ValueError("Wave Gate must contain exactly three to six children")
-    child_fields = {
-        "assignmentId",
-        "workDefinitionRoot",
-        "deliveryClass",
-        "responsibilitySlices",
-        "dependsOn",
-        "terminal",
-    }
-    child_ids: list[str] = []
-    for child in children:
-        _strict_object(child, child_fields, "child")
-        assignment_id = child["assignmentId"]
-        if not isinstance(assignment_id, str) or not assignment_id:
-            raise ValueError("child.assignmentId must be a non-empty string")
-        child_ids.append(assignment_id)
-        _root(child["workDefinitionRoot"], "child.workDefinitionRoot")
-        if child["deliveryClass"] not in FAMILY_DELIVERY_CLASSES:
-            raise ValueError("child.deliveryClass is unsupported")
-        _sorted_unique_strings(
-            child["responsibilitySlices"], "child.responsibilitySlices"
-        )
-        _sorted_unique_strings(child["dependsOn"], "child.dependsOn", allow_empty=True)
-    expected_ids = sorted(set(child_ids), key=lambda row: row.encode("utf-8"))
-    if child_ids != expected_ids:
-        raise ValueError("child membership must be sorted and duplicate-free")
-    members = set(child_ids)
-    graph: dict[str, list[str]] = {}
-    for child in children:
-        assignment_id = child["assignmentId"]
-        dependencies = child["dependsOn"]
-        if assignment_id in dependencies:
-            raise ValueError("child cannot depend on itself")
-        orphan = next((row for row in dependencies if row not in members), None)
-        if orphan:
-            raise ValueError(f"child dependency is not a Wave member: {orphan}")
-        graph[assignment_id] = dependencies
-        if child["terminal"] is not None:
-            _validate_family_terminal(child["terminal"], child)
-
-    visiting: set[str] = set()
-    visited: set[str] = set()
-
-    def visit(node: str) -> None:
-        if node in visiting:
-            raise ValueError("child dependency graph contains a cycle")
-        if node in visited:
-            return
-        visiting.add(node)
-        for dependency in graph[node]:
-            visit(dependency)
-        visiting.remove(node)
-        visited.add(node)
-
-    for child_id in child_ids:
-        visit(child_id)
-
-    acceptance = state["acceptance"]
-    if not isinstance(acceptance, list) or not acceptance:
-        raise ValueError("family acceptance matrix must be non-empty")
-    acceptance_ids: list[str] = []
-    for row in acceptance:
-        _strict_object(
-            row, {"acceptanceId", "status", "evidenceRoots"}, "acceptance row"
-        )
-        acceptance_id = row["acceptanceId"]
-        if not isinstance(acceptance_id, str) or not acceptance_id:
-            raise ValueError("acceptance.acceptanceId must be a non-empty string")
-        acceptance_ids.append(acceptance_id)
-        if row["status"] not in FAMILY_ACCEPTANCE_STATES:
-            raise ValueError("acceptance.status is unsupported")
-        roots = _sorted_unique_strings(
-            row["evidenceRoots"], "acceptance.evidenceRoots", allow_empty=True
-        )
-        for root in roots:
-            _root(root, "acceptance.evidenceRoots")
-        if row["status"] == "missing" and roots:
-            raise ValueError("missing acceptance cannot claim evidence")
-        if row["status"] != "missing" and not roots:
-            raise ValueError(
-                f"{row['status']} acceptance requires exact evidence roots"
-            )
-    if acceptance_ids != sorted(
-        set(acceptance_ids), key=lambda row: row.encode("utf-8")
-    ):
-        raise ValueError("acceptance rows must be sorted and duplicate-free")
-    declared = _root(state["stateRoot"], "stateRoot")
-    if declared != _family_state_root(state):
-        raise ValueError("Initiative-family state root does not verify")
-    return state
-
-
-def create_family_state(blueprint: Any) -> dict[str, Any]:
-    """Create the first immutable state from one bounded Wave blueprint."""
-
-    value = _strict_object(
-        blueprint,
-        {"schema", "initiative", "wave", "children", "acceptanceIds"},
-        "Initiative-family blueprint",
-    )
-    if value["schema"] != FAMILY_BLUEPRINT_SCHEMA:
-        raise ValueError(f"family blueprint schema must be {FAMILY_BLUEPRINT_SCHEMA}")
-    initiative = _strict_object(
-        value["initiative"], {"initiativeId", "versionRoot"}, "initiative"
-    )
-    wave = _strict_object(
-        value["wave"], {"waveId", "ordinal", "gateAssignmentId"}, "wave"
-    )
-    children = value["children"]
-    if not isinstance(children, list):
-        raise ValueError("blueprint.children must be an array")
-    normalized_children = [
-        {
-            **_strict_object(
-                child,
-                {
-                    "assignmentId",
-                    "workDefinitionRoot",
-                    "deliveryClass",
-                    "responsibilitySlices",
-                    "dependsOn",
-                },
-                "blueprint child",
-            ),
-            "terminal": None,
-        }
-        for child in children
-    ]
-    acceptance_ids = _sorted_unique_strings(
-        value["acceptanceIds"], "blueprint.acceptanceIds"
-    )
-    state: dict[str, Any] = {
-        "schema": FAMILY_STATE_SCHEMA,
-        "initiative": {**initiative, "role": "inert-parent"},
-        "wave": {**wave, "gateState": "terminal"},
-        "children": normalized_children,
-        "acceptance": [
-            {"acceptanceId": row, "status": "missing", "evidenceRoots": []}
-            for row in acceptance_ids
-        ],
-        "previousStateRoot": "",
-    }
-    state["stateRoot"] = _family_state_root(state)
-    return validate_family_state(state)
-
-
-def transition_family_state(current: Any, transition: Any) -> dict[str, Any]:
-    """Append one exact-root family transition without mutating prior state."""
-
-    prior = validate_family_state(current)
-    update = _strict_object(
-        transition,
-        {
-            "schema",
-            "expectedStateRoot",
-            "terminalUpdates",
-            "acceptanceUpdates",
-        },
-        "Initiative-family transition",
-    )
-    if update["schema"] != FAMILY_TRANSITION_SCHEMA:
-        raise ValueError(f"family transition schema must be {FAMILY_TRANSITION_SCHEMA}")
-    expected = _root(update["expectedStateRoot"], "expectedStateRoot")
-    if expected != prior["stateRoot"]:
-        raise ValueError("Initiative-family state changed before transition")
-    terminal_updates = update["terminalUpdates"]
-    acceptance_updates = update["acceptanceUpdates"]
-    if not isinstance(terminal_updates, list) or not isinstance(
-        acceptance_updates, list
-    ):
-        raise ValueError("family transition updates must be arrays")
-    if not terminal_updates and not acceptance_updates:
-        raise ValueError("family transition must change at least one field")
-    result = json.loads(canonical_json(prior))
-    children = {row["assignmentId"]: row for row in result["children"]}
-    semantic_change = bool(terminal_updates)
-    terminal_ids: list[str] = []
-    for row in terminal_updates:
-        _strict_object(row, {"assignmentId", "terminal"}, "terminal update")
-        assignment_id = row["assignmentId"]
-        terminal_ids.append(assignment_id)
-        if assignment_id not in children:
-            raise ValueError(f"terminal update names an orphan child: {assignment_id}")
-        child = children[assignment_id]
-        if child["terminal"] is not None:
-            raise ValueError("terminal child state is immutable")
-        _validate_family_terminal(row["terminal"], child)
-        child["terminal"] = row["terminal"]
-    if terminal_ids != sorted(set(terminal_ids), key=lambda row: row.encode("utf-8")):
-        raise ValueError("terminal updates must be sorted and duplicate-free")
-    for assignment_id in terminal_ids:
-        child = children[assignment_id]
-        if any(children[row]["terminal"] is None for row in child["dependsOn"]):
-            raise ValueError("child cannot terminate before all dependencies")
-
-    coverage = {row["acceptanceId"]: row for row in result["acceptance"]}
-    coverage_ids: list[str] = []
-    allowed = {
-        "missing": {"missing", "partial", "proved", "invalidated"},
-        "partial": {"partial", "proved", "invalidated"},
-        "proved": {"proved", "invalidated"},
-        "invalidated": {"invalidated"},
-    }
-    for row in acceptance_updates:
-        _strict_object(
-            row, {"acceptanceId", "status", "evidenceRoots"}, "acceptance update"
-        )
-        acceptance_id = row["acceptanceId"]
-        coverage_ids.append(acceptance_id)
-        if acceptance_id not in coverage:
-            raise ValueError(
-                f"acceptance update names an unknown item: {acceptance_id}"
-            )
-        current_coverage = coverage[acceptance_id]
-        if row["status"] not in allowed[current_coverage["status"]]:
-            raise ValueError("acceptance coverage transition is invalid")
-        if row != current_coverage:
-            semantic_change = True
-        current_coverage.clear()
-        current_coverage.update(row)
-    if coverage_ids != sorted(set(coverage_ids), key=lambda row: row.encode("utf-8")):
-        raise ValueError("acceptance updates must be sorted and duplicate-free")
-    if not semantic_change:
-        raise ValueError("family transition produced no semantic change")
-    result["previousStateRoot"] = prior["stateRoot"]
-    result.pop("stateRoot", None)
-    result["stateRoot"] = _family_state_root(result)
-    return validate_family_state(result)
-
-
-def verify_family_state(value: Any) -> dict[str, Any]:
-    state = validate_family_state(value)
-    terminal_counts = {
-        terminal: sum(
-            1
-            for child in state["children"]
-            if (child.get("terminal") or {}).get("state") == terminal
-        )
-        for terminal in FAMILY_TERMINAL_STATES
-    }
-    return {
-        "schema": "kungfu.work-control.initiative-family-verification/v1",
-        "ok": True,
-        "stateRoot": state["stateRoot"],
-        "previousStateRoot": state["previousStateRoot"],
-        "waveGateTerminal": True,
-        "parentInert": True,
-        "childCount": len(state["children"]),
-        "terminalCounts": terminal_counts,
-        "waveDrained": sum(terminal_counts.values()) == len(state["children"]),
-        "acceptance": [
-            {
-                "acceptanceId": row["acceptanceId"],
-                "status": row["status"],
-            }
-            for row in state["acceptance"]
-        ],
-    }
-
-
-def _family_v2_state_root(value: Mapping[str, Any]) -> str:
-    preimage = dict(value)
-    preimage.pop("stateRoot", None)
-    return semantic_root(preimage)
-
-
-def _family_v2_binding_root(value: Mapping[str, Any]) -> str:
-    preimage = dict(value)
-    preimage.pop("bindingRoot", None)
-    return semantic_root(preimage)
-
-
-def _validate_family_v2_reference(
-    value: Any,
-    *,
-    expected_kind: str,
-    expected_fact_world: str,
-    expected_cut_root: str,
-    expected_status: str,
-    label: str,
-) -> dict[str, Any]:
-    reference = _strict_object(
-        value,
-        {"kind", "identity", "root", "factWorld", "cutRoot", "schema", "status"},
-        label,
-    )
-    if reference["kind"] != expected_kind:
-        raise ValueError(f"{label}.kind must be {expected_kind}")
-    if (
-        not isinstance(reference["identity"], str)
-        or not reference["identity"]
-        or not isinstance(reference["schema"], str)
-        or not reference["schema"]
-        or not isinstance(reference["status"], str)
-        or not reference["status"]
-    ):
-        raise ValueError(f"{label} identity, schema, and status are required")
-    _root(reference["root"], f"{label}.root")
-    _root(reference["cutRoot"], f"{label}.cutRoot")
-    if reference["factWorld"] != expected_fact_world:
-        raise ValueError(f"{label} belongs to the wrong fact world")
-    if reference["cutRoot"] != expected_cut_root:
-        raise ValueError(f"{label} belongs to the wrong cut")
-    if expected_kind == "execution-warrant" and reference["status"] != "active":
-        raise ValueError("execution Warrant must be active, not stale or revoked")
-    if expected_kind == "admission-receipt" and reference["status"] != "admitted":
-        raise ValueError("Admission receipt must record an admitted effect")
-    if expected_kind == "publication-failure" and reference["status"] != "visible":
-        raise ValueError("publication failure must remain visible")
-    if reference["status"] != expected_status:
-        raise ValueError(f"{label}.status must be {expected_status}")
-    return reference
-
-
-def _validate_family_v2_optional_reference(
-    value: Any,
-    *,
-    expected_kind: str,
-    expected_fact_world: str,
-    expected_cut_root: str,
-    expected_status: str,
-    label: str,
-) -> dict[str, Any]:
-    optional = _strict_object(value, {"present", "reference"}, label)
-    if not isinstance(optional["present"], bool):
-        raise ValueError(f"{label}.present must be a boolean")
-    if not optional["present"]:
-        if optional["reference"] is not None:
-            raise ValueError(f"{label} must not hide an undeclared reference")
-        return optional
-    if optional["reference"] is None:
-        raise ValueError(f"{label} declares presence without a reference")
-    _validate_family_v2_reference(
-        optional["reference"],
-        expected_kind=expected_kind,
-        expected_fact_world=expected_fact_world,
-        expected_cut_root=expected_cut_root,
-        expected_status=expected_status,
-        label=f"{label}.reference",
-    )
-    return optional
-
-
-def _validate_family_v2_settlement(
-    value: Any,
-    *,
-    expected_fact_world: str,
-    expected_cut_root: str,
-    label: str,
-) -> dict[str, Any]:
-    settlement = _strict_object(
-        value,
-        {"factWorld", "factCutRoot", "references", "publication"},
-        label,
-    )
-    fact_world = settlement["factWorld"]
-    if not isinstance(fact_world, str) or not fact_world:
-        raise ValueError(f"{label}.factWorld must be a non-empty string")
-    fact_cut_root = _root(settlement["factCutRoot"], f"{label}.factCutRoot")
-    if fact_world != expected_fact_world:
-        raise ValueError(f"{label}.factWorld does not match the family binding")
-    if fact_cut_root != expected_cut_root:
-        raise ValueError(f"{label}.factCutRoot does not match the family binding")
-    references = _strict_object(
-        settlement["references"],
-        {
-            "completionClaim",
-            "assessment",
-            "decision",
-            "admissionReceipt",
-            "episode",
-            "projectCut",
-            "deliveryEvidence",
-        },
-        f"{label}.references",
-    )
-    reference_contracts = {
-        "completionClaim": ("completion-claim", "claimed-complete"),
-        "assessment": ("assessment", "fit"),
-        "decision": ("decision", "accepted"),
-        "admissionReceipt": ("admission-receipt", "admitted"),
-        "episode": ("episode", "sealed"),
-        "projectCut": ("project-cut", "settled"),
-        "deliveryEvidence": ("delivery-evidence", "verified"),
-    }
-    for field, (kind, status) in reference_contracts.items():
-        _validate_family_v2_reference(
-            references[field],
-            expected_kind=kind,
-            expected_fact_world=fact_world,
-            expected_cut_root=fact_cut_root,
-            expected_status=status,
-            label=f"{label}.references.{field}",
-        )
-
-    publication = _strict_object(
-        settlement["publication"],
-        {"state", "lagStartedAt", "failure"},
-        f"{label}.publication",
-    )
-    publication_state = publication["state"]
-    if publication_state not in {"published", "pending", "failed"}:
-        raise ValueError(f"{label}.publication.state is unsupported")
-    lag_started_at = publication["lagStartedAt"]
-    if publication_state == "published":
-        if lag_started_at is not None:
-            raise ValueError("published settlement cannot retain publication lag")
-    else:
-        _timestamp(lag_started_at, f"{label}.publication.lagStartedAt")
-    failure = _validate_family_v2_optional_reference(
-        publication["failure"],
-        expected_kind="publication-failure",
-        expected_fact_world=fact_world,
-        expected_cut_root=fact_cut_root,
-        expected_status="visible",
-        label=f"{label}.publication.failure",
-    )
-    if failure["present"] != (publication_state == "failed"):
-        raise ValueError("publication failure visibility does not match its state")
-    return settlement
-
-
-def validate_family_binding_v2(
-    value: Any, v1_state: Mapping[str, Any]
-) -> dict[str, Any]:
-    """Validate caller-supplied semantic bindings against one exact v1 state."""
-
-    projection = validate_family_state(v1_state)
-    manifest = _strict_object(
-        value,
-        {
-            "schema",
-            "v1StateRoot",
-            "factWorld",
-            "factCutRoot",
-            "initiative",
-            "children",
-            "bindingRoot",
-        },
-        "Initiative-family typed binding manifest",
-    )
-    if manifest["schema"] != FAMILY_BINDING_V2_SCHEMA:
-        raise ValueError(f"typed binding schema must be {FAMILY_BINDING_V2_SCHEMA}")
-    if manifest["v1StateRoot"] != projection["stateRoot"]:
-        raise ValueError("typed binding names the wrong v1 predecessor state")
-    fact_world = manifest["factWorld"]
-    if not isinstance(fact_world, str) or not fact_world:
-        raise ValueError("typed binding factWorld must be a non-empty string")
-    fact_cut_root = _root(manifest["factCutRoot"], "typed binding factCutRoot")
-
-    initiative = _strict_object(
-        manifest["initiative"],
-        {"initiativeId", "pursuit", "atlas", "acceptancePolicy"},
-        "typed binding initiative",
-    )
-    if initiative["initiativeId"] != projection["initiative"]["initiativeId"]:
-        raise ValueError("typed binding Initiative identity does not match v1")
-    for field, kind, status in (
-        ("pursuit", "pursuit", "active"),
-        ("atlas", "atlas", "current"),
-        ("acceptancePolicy", "acceptance-policy", "current"),
-    ):
-        _validate_family_v2_reference(
-            initiative[field],
-            expected_kind=kind,
-            expected_fact_world=fact_world,
-            expected_cut_root=fact_cut_root,
-            expected_status=status,
-            label=f"typed binding initiative.{field}",
-        )
-
-    children = manifest["children"]
-    if not isinstance(children, list):
-        raise ValueError("typed binding children must be an array")
-    projected_children = {
-        child["assignmentId"]: child for child in projection["children"]
-    }
-    child_ids: list[str] = []
-    for child in children:
-        bound = _strict_object(
-            child,
-            {
-                "assignmentId",
-                "assignmentState",
-                "workDefinition",
-                "pursuit",
-                "atlas",
-                "executionWarrant",
-                "settlement",
-            },
-            "typed binding child",
-        )
-        assignment_id = bound["assignmentId"]
-        child_ids.append(assignment_id)
-        if assignment_id not in projected_children:
-            raise ValueError(f"typed binding names an orphan child: {assignment_id}")
-        projected = projected_children[assignment_id]
-        reference_contracts = {
-            "assignmentState": (
-                "assignment-state",
-                "terminal" if projected["terminal"] is not None else "active",
-            ),
-            "workDefinition": ("work-definition", "accepted"),
-            "pursuit": ("pursuit", "active"),
-            "atlas": ("atlas", "current"),
-            "executionWarrant": ("execution-warrant", "active"),
-        }
-        for field, (kind, status) in reference_contracts.items():
-            _validate_family_v2_reference(
-                bound[field],
-                expected_kind=kind,
-                expected_fact_world=fact_world,
-                expected_cut_root=fact_cut_root,
-                expected_status=status,
-                label=f"typed binding child.{assignment_id}.{field}",
-            )
-        if bound["assignmentState"]["identity"] != assignment_id:
-            raise ValueError("Assignment-state reference identity does not match child")
-        if bound["workDefinition"]["identity"] != f"{assignment_id}:work-definition":
-            raise ValueError("work-definition reference identity does not match child")
-        if bound["workDefinition"]["root"] != projected["workDefinitionRoot"]:
-            raise ValueError("work-definition reference root does not match v1")
-
-        optional = _strict_object(
-            bound["settlement"], {"present", "value"}, "typed binding settlement"
-        )
-        if not isinstance(optional["present"], bool):
-            raise ValueError("typed binding settlement.present must be a boolean")
-        terminal_state = (projected["terminal"] or {}).get("state")
-        if optional["present"] != (terminal_state == "merged"):
-            raise ValueError(
-                "typed settlement presence must exactly match a merged terminal child"
-            )
-        if optional["present"]:
-            if optional["value"] is None:
-                raise ValueError("typed settlement declares presence without a value")
-            _validate_family_v2_settlement(
-                optional["value"],
-                expected_fact_world=fact_world,
-                expected_cut_root=fact_cut_root,
-                label=f"typed binding child.{assignment_id}.settlement",
-            )
-        elif optional["value"] is not None:
-            raise ValueError("absent typed settlement must not hide a value")
-
-    expected_ids = [child["assignmentId"] for child in projection["children"]]
-    if child_ids != expected_ids:
-        raise ValueError(
-            "typed binding children must exactly match sorted v1 membership"
-        )
-    declared = _root(manifest["bindingRoot"], "typed binding bindingRoot")
-    if declared != _family_v2_binding_root(manifest):
-        raise ValueError("Initiative-family typed binding root does not verify")
-    return manifest
-
-
-def _create_family_state_v2(
-    projection: Mapping[str, Any],
-    bindings: Mapping[str, Any],
-    *,
-    predecessor_state_root: str,
-    previous_state_root: str,
-) -> dict[str, Any]:
-    state: dict[str, Any] = {
-        "schema": FAMILY_STATE_V2_SCHEMA,
-        "predecessorStateRoot": predecessor_state_root,
-        "v1ProjectionRoot": projection["stateRoot"],
-        "v1Projection": json.loads(canonical_json(projection)),
-        "typedBindingRoot": bindings["bindingRoot"],
-        "typedBindings": json.loads(canonical_json(bindings)),
-        "previousStateRoot": previous_state_root,
-    }
-    state["stateRoot"] = _family_v2_state_root(state)
-    return validate_family_state_v2(state)
-
-
-def validate_family_state_v2(value: Any) -> dict[str, Any]:
-    """Validate a fully typed v2 state and its exact v1 projection."""
-
-    state = _strict_object(
-        value,
-        {
-            "schema",
-            "predecessorStateRoot",
-            "v1ProjectionRoot",
-            "v1Projection",
-            "typedBindingRoot",
-            "typedBindings",
-            "previousStateRoot",
-            "stateRoot",
-        },
-        "Initiative-family state v2",
-    )
-    if state["schema"] != FAMILY_STATE_V2_SCHEMA:
-        raise ValueError(f"family state v2 schema must be {FAMILY_STATE_V2_SCHEMA}")
-    _root(state["predecessorStateRoot"], "predecessorStateRoot")
-    previous = str(state["previousStateRoot"] or "")
-    if previous:
-        _root(previous, "previousStateRoot")
-    projection = validate_family_state(state["v1Projection"])
-    if state["v1ProjectionRoot"] != projection["stateRoot"]:
-        raise ValueError("v2 state does not bind its exact v1 projection root")
-    bindings = validate_family_binding_v2(state["typedBindings"], projection)
-    if state["typedBindingRoot"] != bindings["bindingRoot"]:
-        raise ValueError("v2 state does not bind its typed manifest root")
-    declared = _root(state["stateRoot"], "stateRoot")
-    if declared != _family_v2_state_root(state):
-        raise ValueError("Initiative-family state v2 root does not verify")
-    return state
-
-
-def upgrade_family_state_v2(
-    v1_state: Any, typed_binding_manifest: Any
-) -> dict[str, Any]:
-    """Explicitly upgrade one immutable v1 state with caller-supplied bindings."""
-
-    predecessor = validate_family_state(v1_state)
-    bindings = validate_family_binding_v2(typed_binding_manifest, predecessor)
-    successor = _create_family_state_v2(
-        predecessor,
-        bindings,
-        predecessor_state_root=predecessor["stateRoot"],
-        previous_state_root="",
-    )
-    return {
-        "schema": FAMILY_UPGRADE_V2_SCHEMA,
-        "predecessorStateRoot": predecessor["stateRoot"],
-        "typedBindingRoot": bindings["bindingRoot"],
-        "v1ProjectionRoot": predecessor["stateRoot"],
-        "successorStateRoot": successor["stateRoot"],
-        "successorState": successor,
-    }
-
-
-def transition_family_state_v2(current: Any, transition: Any) -> dict[str, Any]:
-    """Advance v2 only with a v1 transition and a complete successor manifest."""
-
-    prior = validate_family_state_v2(current)
-    update = _strict_object(
-        transition,
-        {
-            "schema",
-            "expectedStateRoot",
-            "v1Transition",
-            "typedBindingManifest",
-        },
-        "Initiative-family transition v2",
-    )
-    if update["schema"] != FAMILY_TRANSITION_V2_SCHEMA:
-        raise ValueError(
-            f"family transition v2 schema must be {FAMILY_TRANSITION_V2_SCHEMA}"
-        )
-    expected = _root(update["expectedStateRoot"], "expectedStateRoot")
-    if expected != prior["stateRoot"]:
-        raise ValueError("Initiative-family state v2 changed before transition")
-    successor_projection = transition_family_state(
-        prior["v1Projection"], update["v1Transition"]
-    )
-    bindings = validate_family_binding_v2(
-        update["typedBindingManifest"], successor_projection
-    )
-    return _create_family_state_v2(
-        successor_projection,
-        bindings,
-        predecessor_state_root=prior["predecessorStateRoot"],
-        previous_state_root=prior["stateRoot"],
-    )
-
-
-def project_family_state_v1(value: Any) -> dict[str, Any]:
-    """Return the exact v1-compatible state carried by a typed v2 envelope."""
-
-    state = validate_family_state_v2(value)
-    return json.loads(canonical_json(state["v1Projection"]))
-
-
-def verify_family_state_v2(value: Any) -> dict[str, Any]:
-    """Read v1 as under-typed or verify a complete v2 typed envelope."""
-
-    if isinstance(value, dict) and value.get("schema") == FAMILY_STATE_SCHEMA:
-        state = validate_family_state(value)
-        return {
-            "schema": "kungfu.work-control.initiative-family-verification/v2",
-            "ok": True,
-            "inputSchema": FAMILY_STATE_SCHEMA,
-            "typingState": "under-typed-v1",
-            "stateRoot": state["stateRoot"],
-            "v1ProjectionRoot": state["stateRoot"],
-            "fullyTyped": False,
-            "completionQualified": False,
-            "missingSemanticBindings": [
-                "acceptance-policy",
-                "admission-receipt",
-                "assessment",
-                "assignment-state",
-                "completion-claim",
-                "decision",
-                "delivery-evidence",
-                "episode",
-                "execution-warrant",
-                "atlas",
-                "project-cut",
-                "pursuit",
-                "work-definition",
-            ],
-        }
-    state = validate_family_state_v2(value)
-    projection = state["v1Projection"]
-    merged = [
-        child
-        for child in state["typedBindings"]["children"]
-        if (
-            next(
-                row["terminal"]
-                for row in projection["children"]
-                if row["assignmentId"] == child["assignmentId"]
-            )
-            or {}
-        ).get("state")
-        == "merged"
-    ]
-    publication_states = [
-        child["settlement"]["value"]["publication"]["state"] for child in merged
-    ]
-    v1_verification = verify_family_state(projection)
-    acceptance_proved = all(
-        row["status"] == "proved" for row in projection["acceptance"]
-    )
-    all_children_merged = len(merged) == len(projection["children"])
-    publication_complete = all(row == "published" for row in publication_states)
-    return {
-        "schema": "kungfu.work-control.initiative-family-verification/v2",
-        "ok": True,
-        "inputSchema": FAMILY_STATE_V2_SCHEMA,
-        "typingState": "fully-typed-v2",
-        "stateRoot": state["stateRoot"],
-        "previousStateRoot": state["previousStateRoot"],
-        "predecessorStateRoot": state["predecessorStateRoot"],
-        "v1ProjectionRoot": state["v1ProjectionRoot"],
-        "typedBindingRoot": state["typedBindingRoot"],
-        "fullyTyped": True,
-        "waveDrained": v1_verification["waveDrained"],
-        "mergedSettlementCount": len(merged),
-        "publicationComplete": publication_complete,
-        "pendingPublicationCount": publication_states.count("pending"),
-        "failedPublicationCount": publication_states.count("failed"),
-        "completionQualified": (
-            v1_verification["waveDrained"]
-            and all_children_merged
-            and acceptance_proved
-            and publication_complete
-        ),
-        "projectionMergeIsCompletion": False,
-    }
+# Backward-compatible imports; family behavior is owned by dedicated modules.
+ROOT = assignment_canonical.ROOT
+_ISO_8601 = assignment_canonical._ISO_8601
+_normalized = assignment_canonical._normalized
+canonical_json = assignment_canonical.canonical_json
+semantic_root = assignment_canonical.semantic_root
+_strict_object = assignment_canonical._strict_object
+_sorted_unique_strings = assignment_canonical._sorted_unique_strings
+_timestamp = assignment_canonical._timestamp
+_root = assignment_canonical._root
+
+FAMILY_BLUEPRINT_SCHEMA = initiative_family.FAMILY_BLUEPRINT_SCHEMA
+FAMILY_STATE_SCHEMA = initiative_family.FAMILY_STATE_SCHEMA
+FAMILY_TRANSITION_SCHEMA = initiative_family.FAMILY_TRANSITION_SCHEMA
+FAMILY_DELIVERY_CLASSES = initiative_family.FAMILY_DELIVERY_CLASSES
+FAMILY_TERMINAL_STATES = initiative_family.FAMILY_TERMINAL_STATES
+FAMILY_ACCEPTANCE_STATES = initiative_family.FAMILY_ACCEPTANCE_STATES
+family_contract = initiative_family.family_contract
+validate_family_state = initiative_family.validate_family_state
+create_family_state = initiative_family.create_family_state
+transition_family_state = initiative_family.transition_family_state
+verify_family_state = initiative_family.verify_family_state
+
+FAMILY_CONTRACT_V2_SCHEMA = initiative_family_v2.FAMILY_CONTRACT_V2_SCHEMA
+FAMILY_STATE_V2_SCHEMA = initiative_family_v2.FAMILY_STATE_V2_SCHEMA
+FAMILY_TRANSITION_V2_SCHEMA = initiative_family_v2.FAMILY_TRANSITION_V2_SCHEMA
+FAMILY_BINDING_V2_SCHEMA = initiative_family_v2.FAMILY_BINDING_V2_SCHEMA
+FAMILY_UPGRADE_V2_SCHEMA = initiative_family_v2.FAMILY_UPGRADE_V2_SCHEMA
+FAMILY_V2_REFERENCE_KINDS = initiative_family_v2.FAMILY_V2_REFERENCE_KINDS
+InitiativeFamilyV1Port = initiative_family_v2.InitiativeFamilyV1Port
+family_contract_v2 = initiative_family_v2.family_contract_v2
+validate_family_binding_v2 = initiative_family_v2.validate_family_binding_v2
+validate_family_state_v2 = initiative_family_v2.validate_family_state_v2
+upgrade_family_state_v2 = initiative_family_v2.upgrade_family_state_v2
+transition_family_state_v2 = initiative_family_v2.transition_family_state_v2
+project_family_state_v1 = initiative_family_v2.project_family_state_v1
+verify_family_state_v2 = initiative_family_v2.verify_family_state_v2
 
 
 def source_root(*starts: str | Path) -> Path:
@@ -2374,15 +1340,6 @@ def list_outcome_bindings(workspace_root: str | Path) -> dict[str, Any]:
         "writes": [],
     }
     return {**body, "index_root": semantic_root(body)}
-
-
-def _root(value: Any, field: str, *, optional: bool = False) -> str:
-    text = str(value or "")
-    if optional and not text:
-        return ""
-    if not re.fullmatch(r"sha256:[0-9a-f]{64}", text):
-        raise ValueError(f"{field} must be a sha256 root")
-    return text
 
 
 def _binding_endpoint(
